Fixed the bug that James found involving abstract kinds. This fix was pretty painful.  Basically, we now track two sorts of kind constraints when inferring kinds.  One says that this *must* be the kind, and one says this must be an upper bound on the kind.
A side effect of this change is that you must now explicitly
specify when you want to use abstract kinds in a function
declaration.  This is not great, but it keeps things consistent.

A good side effect of the change is that you can no longer
have the compiler implicitly promote a M or A kind to a boxed
kind which is something that always bugged me when the
user put in an explicit M or A.

There were a lot of other related bugs, the largest being
that we didn't check that kinds were consistent when
unifying two function types.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1570 581db07f-6e62-4d40-9733-d432c4eebb07