* Fixed patches (I hope) * Made it so that -L arguments get put in cyclone_lib_path rather than sent to cpp * Made it so cyclone_lib_path elements (from -B, -L, and default) are getting sent to gcc as -L arguments * Added a command line option -warnboundschecks that emits a warning whenever a bounds check is necessary (as determined by the analysis). * Fixed a bug in cyclone.cyc: two occurrences of   Filename::chop_extension(output_file==NULL ? s : *output_file); needed to be   output_file==NULL ?  (string_t)Filename::chop_extension(s)  : *output_file;
Problem:  assert uses __FILE__, which is a fully qualified pathname, and so differs from user to user.  (For example, mine are of the form /home/jcheney/src/cyclone/src/file.cyc).  Several libraries (arg, string, bitvec) use assert, and so the genfiles for these are guaranteed to change whenever someone does an update.  This seems irritating.  I notice that GNU cpp lets you do __BASE_FILE__ instead to get just the filename that was specified on the command line, which seems like what we want, but may not be standard for other preprocessors.  Is everyone using GNU cpp?


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1807 581db07f-6e62-4d40-9733-d432c4eebb07