Bug fixes to existential types on structs.  Changed fn.cyc to use a struct instead of a tunion.  The next commit will disallow existential vars on tunion fields.
Here's an example file that compiles.  The initialization and witness
types for s3 make me smirk -- one of those moments where the compiler
works exactly like it should and it's code you wrote, but it still
surprises you.

struct Foo { <`a,`b> `a x; `a y; `b z; };

void f() {
  struct Foo s1 = Foo(3,4,"blah");
  struct Foo s2 = Foo("blah","blaz",17);
  struct Foo s3 = Foo(&s1, &s2, &s3);
  struct Foo s4;
  s4 = s3;
  let Foo{<`c,`d> f1,f2,f3} = s3;
  f1==f2;
}


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1840 581db07f-6e62-4d40-9733-d432c4eebb07