* Fixed a syntax error in check_known_subscript_notnull.  Apparently this code is seldom exercised.  As a result, every genfile changed...
* After Dan redid the rgnorder stuff, I ran into uncaught exceptions
resulting from dictionary lookup failures in rgnorder.cyc.  I have
wrapped these inside if(Dict::member) { ... }, and if the membership
test fails I do nothing.  I'm not sure this is the right behavior.  Dan?

* Updated the uses of unsafe_cast in xmlmarshal so that they typecheck
again.

* Minor error message improvement in evexp

* In tcpat, changed the kindbounds checking for existential unpack
forms so that you no longer get an error when the kind of an unpacked
variable is not BoxKind.  Instead, it now checks that the kind of the
pattern variable is greater than or equal to that of the corresponding
bound struct variable.  I'm probably not catching all the possible cases,
but it does get the common case, where the bound variable is constrained
equal to something and the pattern variable is equal to another kind
or unconstrained.  So it infers the most general (least) acceptable kind
if you leave the kind off.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1889 581db07f-6e62-4d40-9733-d432c4eebb07