Okay, I hacked in support for sizeof (and offsetof) in constant positions, particularly pointer and array bounds.  Basically we're careful about saying "okay, this is a constant for C, but I know not which constant it is".  I spent about 3 minutes testing this feature -- the basics work.
There are places where "not knowing the constant" is unacceptable.
One that won't change is tuple projection.  Out of "hacker laziness",
I also forbid sizeof or offset of in:
  * bit-field sizes
  * explicit enum tags
  * switch "C" cases

Finall, when a comprehension (it's always the damn comprehensions)
isn't under a new, it's bound needs to be a value we know because we
expand out the comprehension at compile-time.  But we're okay if the
body of the comprehension is 0.  Sigh.

Who knows what sort of damage I did to tcGenrep.

Toc::enumfields_to_c is now a no-op -- this should be fine because the
field tags are constants.

Also, we can't check if you try to make an array too big such as:
int x[sizeof(int)-5];
gcc catches this for us though.

It's a little hard to explain what the rule for "are these two
constants equal" is when you use sizeof or offsetof.  But I do know
what the rule is (modulo bugs).

I did update_all_archs and it took less than half the time it used
to.  I assume that's a good thing.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1939 581db07f-6e62-4d40-9733-d432c4eebb07