* Trevor, you can now subtract pointers.
* Mike, you no longer have to initialize zero-terminated arrays
within functions (the compiler puts in the last 0 for you.)

* You can also malloc a zero-terminated array, but this turns
into a calloc and you get a warning.

* Fixed the error messages somewhat, but still not completely
satisfactory.  What we really need to do is write a routine
which takes two types and explains why they are incompatible
in a given context.  At least we no longer say that
"char and char do not unify".  :-)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1998 581db07f-6e62-4d40-9733-d432c4eebb07