Some little stuff to make using tag_t types (and other stuff) a bit more convenient:
* Evexp works through casts to tag_t types, but rejects negative
numbers.  (Perhaps it should allow it and just create a large positive
number.)

* Parser accepts tag_t (without <t>) and generates a fresh Evar of
  IntKind.

* Parser accepts x<_> as a pattern, replacing _ with a fresh Tvar of
  IntKind.

* Allow a pattern to match an expression so long as coerce_assign
  allows the expression to the pattern's type.  (We previously only
  allowed unifiable types.)

That gets my favorite function down to this:

struct T<`a> { <`i::I>
   tag_t<`i> sz;
  `a @{`i}   arr;
};
struct T<`a> doubleSz(struct T<`a> oldarr) {
  let T{<`n> sz,elts} = oldarr;
  tag_t x = 0;
  if(sz<=x)
    return oldarr;
  let newsz<_> = sz*2;
  return T(newsz, new {for j < newsz: j < sz ? elts[j] : elts[x]});
}

Remaining things to do that might be helpful:
1. allow leaving off tvar names in the unpack (i.e. elide the <`n> and
generate for the user)
2. have the flow analysis maintain IntKind relations at program points,
   for example (`0 < `n).  Doing so would let us replace the two uses
   of x with (tag_t)0.  The cast is necessary to get a tag_t type and
   not an int type.  I fear using context to "guess it's a tag" would
   be confusing, but we have some ideas on rules for doing it.

It's tempting to have "tag_t x = e" mean an existential unpack of the
int expression e.  But that's really not uniform with the rest of the
type system (filling in elided type params with Evars not bound
variables), so I'm inclined to stick with the "let x<_>" form.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2179 581db07f-6e62-4d40-9733-d432c4eebb07