A number of nice additions thanks to my travel time.  These are in increasing order of importance.  Make sure you look at the last change, as it's the most important:
* Made a distinction between long and int, since C does.  However,
internally we treat them as equal.  In the long run, I plan to make
long the boxed type and int not a boxed type.  This will make us
more compatible with 64-bit architectures.

* Added "as" patterns.  You can write "x as <pat>" just as in ML to
match a value against a pattern and bind the value to x.  You can also
write "*x as <pat>" to bind the LHS to the variable x.

* Eliminated some redundant tests for switch's on tunions that
involve a mix of value-carrying and non-value-carrying constructors.

* Added downward nested functions.  You can now write something like:

   int sum(list_t<int> x) {
     int counter = 0;
     void add_elt(int i) {
       counter += i;
     }
     List::iter(add_elt,x);
     return counter;
   }

Note that add_elt contains a reference to a free variable here
(counter).  In general, nested functions, as you might hope, can refer
to free type and value variables that are in scope.

The way this type-checks is that in the example above, add_elt is
given the type "void (@`r)(int i)" where `r is the name of the region
in the block where add_elt was declared.  To call add_elt, `r still
has to be in scope.  In turn, this prevents you from accessing the
counter after it's deallocated.  (One way to think of this is that `r
is a bound on all of the regions that were in scope, so if `r is live,
then any variable living in a region that was in scope at the point of
definition of the function must also be live.)  Of course, this prevents
you from writing upwards-escaping functions (e.g., compose, curry,
uncurry, etc.)  But that's what the fn library is good for.

The translation is trivial since GCC supports nested functions like
this.  (Of course, it won't be trivial with a different C compiler...)

One minor thing to watch out for with this change: it used to be that
a function parameter e.g., f in "void foo(int f(int))" would be
assumed to be in the heap region.  That is, the above would expand to
"void foo(int (@`H f)(int))" but now it expands to "void foo(int (@`r
f)(int))" where `r is a fresh region variable.  This didn't actually
break very much in the compiler, except for the dict, set, hashtable,
and fn libraries.  Each of these defined a data structure (e.g.,
struct Set) which contains a function pointer.  By default, those
function pointers are assumed to be in the heap.  But the functions to
build those data structures became region polymorphic under the new
default, so we got a type-mismatch (trying to put @`r function pointer
into a data structure expecting @`H.)  The solution is to change the
prototypes for these functions so that they explicitly take a `H
function pointer.  (Actually, the right thing to do is parameterize the
data structure by another region and use that on the function pointer
type.  I just didn't feel like doing this...)

4. The biggest change of all is that I've changed the interpretation
of ? pointer types and added a new pointer type, ?-.

Values of ? type now only take two words (the curr and last_plus_one)
fields, but only support non-negative subscripting and pointer
arithmetic (e.g., x++, x + e, x[e], x += e where e is non-negative.)
If you use a negative index, then you'll get an array bounds error at
the point of dereference.

Values of ?- type are just like our old fat pointers -- three words
and arbitrary subscripting and pointer arithmetic.

I chose the two-word version for the default because I felt like this
was what was needed in most situations.  This is validated by the fact
that throughout the compiler, the only major changes needed were to
printf.cyc, scanf.cyc, and the bison sources.  These were all ported
from C code and used a lot of pointer arithmetic.  In the rest of the
comiler, we tended to use subscripting from the base instead of
pointer arithmetic, hence there was no need for negative offsets.  As
a result, the compiler is a little smaller (less code for manipulating
2 words than 3 words).  I've yet to figure out if this is any faster.

I've ported the files in the top-level of tests but none of the others
(e.g., benchmarks, shootout, etc.)  Other code (e.g., the benchmarks
used in the PLDI paper, Medianet, etc.) also needs to be ported.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2264 581db07f-6e62-4d40-9733-d432c4eebb07