Fixed an embarassing bug in the uniqueness flow analysis (actually got the polarity of the lessthan stuff wrong for consumed variables; I'm surprised it worked before!).  Also added some warnings in the case that you clobber a unique variable without freeing it (still need to add warnings for the case that you don't free something before it goes out of scope).  Finally, added a runtime function Core::ufree for freeing unique pointers.  There may be a problem with this, depending on the implementation of the collector, since it has type
extern void ufree(`a::A ?-`U ptr);

The C function frees that base pointer, but if you did something like

int ?`U x = new { for i < 3; i };
int *`U y = x + 1;
Core::ufree(y);

Then the GC_free routine would be called with a pointer into the
middle of the allocated object.  It could do the right or the wrong
thing in this case, I'm not sure.  If this is a problem, we probably
have to forbid pointer arithmetic on unique pointers, to prevent
losing the base.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2267 581db07f-6e62-4d40-9733-d432c4eebb07