The tag_t<`i> type is now useful again.  In particular, see tests/shootout/cyc-matrix.cyc which is now a faithful port of the original matrix.gcc with (a) no tagging, (b) no bounds checks.  Rather, everything is done within the type system. Obviously, we now get the same performance for matrix multiply :-) (Oh yes, I even region-converted it...)
This involved tricking malloc/calloc and new {for i < e1 : e}
into returning T@{valueof(`i)} when the type of the number
of elements was tag_t<`i>.  It also involved beefing up the
relations tracked by the flow analysis (but not by much) and
the toc.cyc translation so that the bounds checks could be
eliminated.  The way I did this is a little more robust than the
way it was done previously, so it should work a bit better.  In
particular, other than change the prototype for the
matrix multiply routine, and change the index variables
from int to unsigned int, I didn't have to change anything
to get the code to compile without any bounds checks.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2323 581db07f-6e62-4d40-9733-d432c4eebb07