* changed syntax for region and alias declarations, as per Trevor's suggestion.  In particular, they are treated in both the abstract and concrete syntax as declarations.  So, the right thing to do is write:
{ region r;    ... }

{ alias <`r> x = e;    ... }

Just like let and other declarations, if you don't group things, then
you get a sort of longest scope over the following statements.  (i.e.,
braces are rarely necessary).  Modified the compiler to use the new
syntax.

* modified the compiler so that it supports both fine-grained and
coarse-grained regions during compilation.  Coarse-grained regions
basically allocate one region for per function during type-checking,
flow analysis, and translation.  The fine-grained regions (the old
code) allocates regions on a much more fine-grained level.  For
instance, each variable declaration causes a new region to be
allocated in type-checking.  As expected, the coarse-grained regions
use more space, but it seems every so slightly faster than the
fine-grained regions.  So, by default, I've left things as
coarse-grained.  The #define that controls this is in tcenv.h (for now
good reason.)  But you'll see macro's of the form: TEMP_RGN(te, r)
which either expand into: region_t r = te->rgn; or else region r; (the
former being a re-use of an existing region and the latter being a
fresh region declaration.)

* allow region_t<`r1> to be cast to a region_t<`r2> when `r1 outlives
`r2.

* added a function Core::mkfat with prototype:

  `a::A ?`r  mkfat(`a::A @{valueof(`n)}`r, Core::sizeof_t<`a>, tag_t<`n>)

which makes a fat pointer out of a thin pointer, its element size, and
the number of elements.  This is useful when you need to do a
subscript on a thin pointer whose bound is given by some value of
tag_t type.  For instance:

  int i;
  tag_t<`n> n;
  int @{valueof(`n)} arr;

  arr[i]

will not type-check because we have no way to ensure that i is in
bounds and the compiler doesn't know where to find the bound for arr.
You can get around this by doing:

  mkfat(arr,sizeof(int),n)[i]

The definition of mkfat is given in core.cyc using the extern "C
include" facility.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2327 581db07f-6e62-4d40-9733-d432c4eebb07