Changed cycswap(x,y) to be x :=: y as discussed earlier.
Changed malloc() and new() to be a bit smarter about inferring
what region to allocate in.  In particular, if tcMalloc
or tcNew gets a topt, indicating that the type to be returned
is given by the context, we base the region on that type, using
the unique region or the heap region.  As a consequence of this
change, I was able switch to using new for all unique pointer
allocations in my test files.

I also changed the shootout benchmark heapsort.cyc to use
unique pointers and our Core::ufree.

Question: at the moment, free() is a no-op defined in buildlib.
This is for porting purposes.  Should we change our Core::ufree()
to be free() instead, or leave it as it is?


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2328 581db07f-6e62-4d40-9733-d432c4eebb07