Changes to get things to build under the latest Cygwin (1.3.15):
* Fixed a bug in the parsing of macro arguments in buildlib -- the
parser required spaces between arguments.  Now a space is optional.

* Fixed a bug in the pretty-printer and dumper:  A statement expression
must be of the form "S;e;".  When pretty-printing/dumping a sequence
statement "S1;S2", we were checking to see if S2 has declarations and
if so, printing "S1;{S2}".  If S2 happens to be of the form S3;e;
then we'd get "S1;S3;e;" printing as "S1;{S3;e;}" which is not correct
for statement expressions.  Now we detect this situation and rewrite
as "S1;({S3;e;});".

* Apparently, fseeko and ftello are not defined in this version of
Cygwin, so I had to #ifdef them out in libc.cys.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2395 581db07f-6e62-4d40-9733-d432c4eebb07