Hmm, many many files changed once again...  I changed the flow analysis, so I don't think I'm the cause for this new behavior (see below).
Basically, I extended the absRval Aggregate() type
to include an additional last_field_no field that indicates the
last field that was written to the Aggregate, if it is a tagged union;
this is -1 when not known.  We use this information to rule out silly
programs like

  @tagged union foo { int x; int y };

  union foo f = {.x = 1};
  printf("%d",f.y);

The flow analysis also considers TagCheck_e's (updating the flow
if it knows the answer), and assignments.  The next step would
be to add a new kind of annotation for tag check elimination and
to create an analogue of anal_derefL for tagchecking.

I additionally had to change the typ_to_absrval initialization rules
so that tagged unions, even when their contents are bits-only, are not
considered initialized by default.  This is sensible, because you'll
get a match exception if you try to access an uninitialized tagged
union at runtime.

I haven't made it so uniqueness (of tagged union fields) properly
interacts with this stuff.  Will do so next.

Finally, I added some tests to cyctest.cyc for this stuff.  Right
now, the one for swapping fails, implying that Greg's last check-in
wasn't quite right.  Perhaps this is the cause for all the genfile
changes?


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2493 581db07f-6e62-4d40-9733-d432c4eebb07