Added reference-counted pointers to Cyclone.  See tests/refcnt_test.cyc for some examples.  Here's what happened:
1) Add a new region `RC that denotes the reference-counted region. This
region has kind TR. which indicates that pointers into it cannot be
aliased (implicitly) and cannot be freed.� Finally, just like unique
pointers, you can drop reference-counted pointers without decrementing
the count; this just means you'll never be able to free the pointer
explicitly, but the GC will get it (assuming it's actually unreachable).

As far as region ordering goes, we have RC >= H but RC is not >= U.
Furthermore, because we can now have concrete values of TR kind,
it is no longer sound for `U or `H to outlive UR or TR kinded pointers.
For that matter, `RC will not outlive arbitrary TR-kinded type
variables (since they could be "bound" to R or UR kinded variables).

2) We additionally define the constant Core::refcount_region, having
type region_t<`RC>, for creating reference-counted pointers.� The caveat
here is that when you allocate something in this region, an extra word
will be prepended to your data, which contains the reference count,
initialized to 1.

As with unique pointers, no pointer arithmetic is allowed, for similar
reasons: it can occlude where the "head" of the object is, and thus make
it impossible to find the hidden reference count.� The reference count
can be accessed by the user via the routine refPtrCount:

� $(int,`a ?`RC) refptr_count(`a::TA ?`RC ptr);

The constant NULL is allowed to have type `a::TA ?`RC, and its reference
count is always 0.

3) Since implicit aliasing is not allowed, aliases can be created
explicitly using the routine aliasRefPtr:

� $(`a ?`RC,`a ?`RC) alias_refptr(`a::TA ?`RC ptr);

This routine will consume the pointer that was passed in, and return two
copies of the pointer.� Furthermore, the reference count will be
updated.

4) To explicitly drop an alias, we use the routine freeRefPtr:

� void drop_refptr(`a::TA ?`RC ptr);

In the case that the provided reference count is 1 (and is thus dropped
to zero), then the provided pointer would be freed.� Note that the flow
analysis will consume the passed pointer (as is always the case for
function arguments), so you won't be able to use it afterwards.

5) Just like unique pointers, reference-counted pointers can be stored
in normal, aliasable datastructures, and accessed using swap.� Because
NULL is a a::TA ?`RC pointer, we can always cheaply construct a pointer
to swap in.� Also, the alias<> construct should work to create temporary
(non-counted) aliases of such a pointer.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2550 581db07f-6e62-4d40-9733-d432c4eebb07