* Expanded the tests/foo example so that it has more than one kind of semantic value, just to make it a better example.
* Added support for writing malloc(x) when x has type
sizeof_t<T> == tag_t<valueof_t<sizeof(T)>>.  Useful for polymorphic
code that needs to allocate some value of abstract type.

* Added support for structs whose last element is of abstract kind.
For instance:

  struct Node<`a::A> { struct Node<`a> *next; `a value; };

This type has abstract kind, so you can only manipulate pointers
to such a value.  Note, however, that if you instantiate `a with
a type of non-abstract kind (e.g., double) then the resulting
type has Mem kind.  This allows you to write some more polymorphic
code.  A good example is in tests/flatlist.cyc.  The most impressive
example in there is map because you must pass in the size of
the resulting nodes (so we can allocate them) and a function
which can be used to initialize the nodes.

There are still some bugs in this feature.  In particular, the
compiler will choke if it sees:

  struct Foo<`a::A> { struct Foo<`a::A> x; };

and end up with a stack overflow.  Also, we do not yet properly
support existentially quantified structs such as:

  struct Bar {<`a::A> int x; `a y;}

which is what's really needed to do good object encodings.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2559 581db07f-6e62-4d40-9733-d432c4eebb07