Trevor, you forgot to check in aprof_newgraph.cyc, so I had to comment out the code that depended on it, including the references in the Makefile, aprof.cyc, and aprof_event.h, but it should be easy to undo these changes when you check it in.
* Changed aprof_graph.cyc to use hash tables instead of linked lists
for storing the dyn_regions and dyn_live lists.  Very large
improvement in performance :-).  Needed this to deal with the amon.out
files generated when running the compiler which are huge.  In
particular, with fine-grained regions, aprof took more than my lunch
time to crunch through just half of the input.  Now it does it in less
than a few minutes.  A nice testament to having good libraries sitting
around.

* The flag -warnregioncoerce can be used to warn when you are
implicitly coercing from one region to another (e.g., from int@`H to
int@`r).  The goal is to help avoid leaks where you heap or unique
allocate something and then pretend it's in a region.  It's not all
that useful since there's a lot of region promotion going on in the
compiler.  I think it would help more if we introduced a new static
region (`S) for the statically- allocated data which was distinct from
the heap and allowed it to implicitly outlive everything, but
otherwise warn when coercing a heap or unique pointer to another
region.

* Fixed a few typos in the documentation (thanks to Mujtaba Ali.)

* Changed tcCast so that it returns the type that you're casting to
upon failure to try to improve error messages.  Doesn't actually
improve error messages much -- need to work on this and come up with a
more systematic way of explaining what went wrong.

* Turned off COARSE_REGIONS by default except within the flow
analysis.  (Turning it off there causes a couple of type-errors, due
to recent changes that I failed to check with this flag turned off.)
This might be a tad slower, but it seems to reduce the working set
size.  I think that if we actually got the fine-grained regions going
for the flow analysis again, then we could reduce the working- set
size further.

* Added support for building cyclone_pg so that we can profile the
compiler itself (but not including the collector yet...)  Found out we
were spending around 57% of our time in the parser.  Figured it might
be due to switching back to a dynamically sized array for the semantic
values and locations (and thus we're now doing null and heavier-weight
array bounds checks.)

* Simplified Position::seg_t to just a single integer.  It used to be
a start and end position, but we don't actually use the end position
anywhere and this saves a fair amount of allocation (about 0.5 MB when
compiling parse_tab.cyc, which was about 5% of the total heap
allocation and the third biggest allocating site in the compiler,
after Absyn::new_exp and Absyn::const_exp.)  This might break the
porting tool -- need to try it out.

* Did a little common-sub-expression elimination on parse.y to avoid
projecting the same component out of yyvs or yyls twice (e.g., having
$n more than once in a production).  This seemed to reduce the
percentage time spent in the parser from 57% to about 50%.

* Modified bison so that at the beginning of each action, it declares
a couple of pointers (one for yyvs and one for yyls) of type @{n}
where n is the number of semantic values that production expects on
the stack.  It initializes them with yyvs+stackoffset-n and
yyls+stackoffset-n as appropriate.  Then, when projecting a value out
of the stack, we use this pointer variable.  Thus, we eliminate all
but one null and bounds check per production which should be better
than even the statically sized arrays.  Sadly, this had no effect on
performance according to the profiler.  Still, it makes me feel warm
and fuzzy to be able to do this.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2595 581db07f-6e62-4d40-9733-d432c4eebb07