Lots of little hacks to eliminate heap allocation for ephemeral data, primarily in the parser.  For instance, I got rid of the need to use a list of type specifiers and instead, we use an unboxed product with bitfields.  As another example, we use unboxed lists (flat lists where the head has abstract type but comes last in a cons cell, so the components of the list do not have to be boxed) for declarator lists. Parameterized the parser (and the YYSTYPE tagged union) by a region which can be used for allocating data local to the parser. Declarators and other such do-hickey's are now allocated there.
Elsewhere, I statically allocated as much as I could (e.g., frequently
used constants such as 0 and 1.)  Got rid of UnknownId_e and
UnknownCall_e since we did a lot of allocation replacing these in the
type-checker.  Now, the parser defaults to Var_e and FnCall_e for
these, but the type-checker might change them when these aren't right.

Used the topt where possible to avoid constructing a new type in
tcexp.cyc.

Got rid of a lot of opt_t's (e.g., opt_t<type_t>) by using
possibly-null pointer types instead (e.g., datatype Type *) which
avoids the need for an option cell.

A few fixes in toc.cyc to deal better with structs whose last members
are abstract.

In the control-flow analysis and toc, we don't substitute for region
parameters or integer type parameters where possible.  This cuts down
on a *lot* of allocation, particularly in the flow analysis.

Avoided allocating a list in the lexer when building a qvar.

Some work to speed up comparison of qvar's since profiling indicated
we were spending a lot of time in there.  Some inlining to speed up
conref manipulation since profiling indicated were spending a lot of
time in it.

Changed aprof so that it doesn't completely die when it finds
an object that is collected which it hasn't seen before.  For
some reason, I'm getting a lot of these...


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2608 581db07f-6e62-4d40-9733-d432c4eebb07