Added lookup_other to dict.h which allows you to use another comparison function than the one associated with the dictionary. Furthermore, there is no requirement that the key passed in has to have the same type as the keys in the dictionary, which is useful when the key lives in a different region.  Took advantage of this is new_control_flow.cyc to avoid allocating a VarRoot just to look it up.
The type-checker lets datatype constructors shadow one another.
That is, if you declare a datatype D1 { C(int); }; and then declare
datatype D2 { C(char); }; the type-checker does not complain and
any subsequent use of C refers to the later definition.  However,
we were not generating unique struct names for the constructors
in the translation to C.  We now incorporate the datatype name
in the struct name.  In particular, the first constructor will
cause us to generate declaration struct Cyc_C_D1_struct { int f1; };
whereas the second one will generate Cyc_C_D2_struct { char f1; };
More generally, the generated struct name includes the namespaces
of both the constructor and the datatype to ensure uniqueness.

This is still not ideal, as these structs can conflict with
user-defined structs.  That is, if at the source level someone
writes "struct C_D1_struct" then we're going to get a conflict.
One way to prevent this would be to collapse the qvar during
translation and prepend a "_" as we do for other temporaries
that are generated.  Then we've reduced the potential conflicts
to those that are only defined within the C namespace.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2614 581db07f-6e62-4d40-9733-d432c4eebb07