Fixed a bug in cyc_include.h which prevented compilation on the Mac. Hopefully it works now.
If you declare a function such as:

  void f() {
    void g();
    g();
  }

then in C, this is equivalent to putting an "extern" in front of the
function declaration for g.  We were getting this wrong and treating g
as if it were a local pointer to a function.  We now treat it as an
extern declaration which, sadly, we do not yet support within
functions.  This turns out to be quite a bit harder to support than I
realized...  While trying to do it, I screwed up the flow analysis
on register declarations, hence the changes to bison.  I fixed the
analysis, but didn't bother undoing my changes to bison.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2620 581db07f-6e62-4d40-9733-d432c4eebb07