Added attribute __noconsume(i)__ to specify that the ith argument of the function it modifies is not consumed by the function.  I've used this to modify some of the reference counting library routines. For example, we used to have
  extern $(int,`a ?`RC) refptr_count(`a::TA ?`RC ptr);

whereas now we have

  extern int refptr_count(`a::TA ?`RC ptr) __attribute__((noconsume(1)));

At the moment there are some restrictions:

1) You cannot declare Cyclone functions with this attribute; it only
applies to runtime system functions.  I plan to add support for noconsume
in Cyclone code shortly.
2) When calling a function with a noconsume parameter, the actual
argument has to be an lval.  This is to make it easier to unconsume
the lval in the flow analysis.  This problem can be fixed, but requires
a more extensive change.  I'm putting off dealing with this until
I need to.

All the genfiles changed again; not sure why.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2634 581db07f-6e62-4d40-9733-d432c4eebb07