Q: why doesn't the standard "make" build an optimized version of the libraries (i.e., with -O)?  I can do "make CFLAGS=-O" I suppose, but it seems better to have this as default.
Got rid of # n on a line when the previous line was labelled
# n-1.  This makes the C files have half as many lines which
seems to make gcc a little happier.

Added structarray.cyc to tests which demonstrates putting an
array of unknown length at the end of a struct.

Got rid of warnings in the parser's C code.  This had to do with
making sure we share instantiations of an abstract struct properly.
I had this right before, but broke it with the support for
arrays and valueof(`i), etc.  Now it works, but it's pretty
complicated...

Changed the way pattern matching is compiled so that we don't
introduce so many temporary variables and we avoid some expensive
copies of structs.  In particular, we used to bind each node in
the pattern to a temporary variable.  For instance, given:

  let &$(a,x,_,_,_,_,_) = x;
  a + x;

we use to generate something like:

  struct tuple7* temp1 = x;
  struct tuple7 temp2;
  int temp3;
  int temp4;
  temp2 = *x;
  temp3 = temp2.f1;
  temp4 = temp2.f2;
  temp3 + temp4;

Note that the assignment "t2 = *x" could be pretty expensive since
we're copying out the whole struct.  Now, we only bind the variables
and not the intermediate nodes, so we generate something like:

  struct tuple4* t1 = x;
  int temp1;
  int temp2;
  temp1 = (*t1).f1;
  temp2 = (*t1).f2;
  temp1 + temp2;

This also simplifies the treatment of the case where variables are
bound to arrays.  Before, I had to jump through some special code to
handle this case, but it falls out naturally now.

However, for deep patterns (e.g., &&&&&$(a,b)), this could be more
expensive since you'll be doing the same projections multiple times
and it's unlikely the C compiler will optimize these away.  The right
thing to do is introduce temps when the path gets too long, but for
now, I haven't bothered, as the compiler actually got smaller and
faster and we never had more than two dereferences in a given pattern
variable definition for the whole compiler.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@2639 581db07f-6e62-4d40-9733-d432c4eebb07